pyproject.toml
[project]
name = "q4-class-11"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.10",
    "py-dotenv>=0.1",
    "rich>=14.1.0",
]
